% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered_fillin2.R
\name{centered_fillin2}
\alias{centered_fillin2}
\title{centered_fillin2}
\usage{
centered_fillin2(method = "both", permute_method = "none", i, j,
  permute = TRUE, ...)
}
\arguments{
\item{method}{determines which linear algebra package to use}

\item{permute_method}{determines how the precision matrix is permuted before use with the chosen linear algebra package}

\item{i}{number of nodes at level 1}

\item{j}{number of children nodes in level 2 per node at level 1}

\item{permute}{determines whether or not to use default row permutation algorithm before matrix factorization to reduce fill-in}

\item{...}{for other params used in centered_precgen2}
}
\value{
the fill-in ratio with respect to the optimal fill-in, and the time taken for the cholesky decomposition
}
\description{
This function returns the ratio of the number of zeros obtained from the cholesky decomposition of the
specified precision matrix over the optimal cholesky fill-in for a centered 2-level gaussian hierarchical model.
If the cholesky decomposition is optimal, the number returned is 1, and if it is sub-optimal the number returned
is greater than 1. The function also returns the time taken for the cholesky decomposition.
}
\details{
A choice of 2 packages is provided: spam and Matrix.
A choice of 3 permute methods is provided: no permutation, random permutation, and depth-first permutation.
}
\examples{
i <- 2
j <- 3
centered_fillin2(method = 'matrix', permute_method = 'random', i = i, j = j)
}
