% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered_gibbs2.R
\name{centered_gibbs2}
\alias{centered_gibbs2}
\title{centered_gibbs2}
\usage{
centered_gibbs2(i, j, ndraws, burnin, flat_prior = TRUE, tau = 1,
  tau_a = 1, tau_b = 1, sigma_2 = 1)
}
\arguments{
\item{i}{number of nodes at level 1}

\item{j}{number of children nodes in level 2 per node at level 1}

\item{ndraws}{number of samples the user wants to have}

\item{burnin}{number of samples to throw away at the start as the gibbs sampler warms up}

\item{flat_prior}{determines whether to use the density with flat prior}

\item{tau}{variance of the root (level 0)}

\item{tau_a}{variance for parameters in level 1}

\item{tau_b}{variance for parameters in level 2}

\item{sigma_2}{variance of the observations}
}
\value{
list of means and the samples
}
\description{
Gibbs sampler for centered 2-level Gaussian hierarchical model according to derived full conditionals
}
\details{
Assumptions:
  - variances are constant for parameters within the same level
  - all observations y_ijk are equal to 0
  - the mean of the root parameter B is mu = 0
  - assume a naive sampler where the variances are not updated
}
\examples{
i <- 2
j <- 3
ndraws <- 10000
burnin <- 1000
centered_gibbs2(i = i, j = j, ndraws = ndraws, burnin = burnin)
}
