% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthfirst.R
\name{depthfirst}
\alias{depthfirst}
\title{depthfirst}
\usage{
depthfirst(i, j, k = NA, levels)
}
\arguments{
\item{i}{number of nodes at level 1}

\item{j}{number of children nodes in level 2 per node at level 1}

\item{k}{number of children nodes in level 3 per node at level 2}

\item{levels}{the number of levels in chosen model}
}
\value{
a sequence of indices which permutes the original matrix
}
\description{
This function returns the sequence which permutes the rows and columns of the precision matrix for a centered 2- or 3-level
gaussian hierarchical model which has been labelled lexicographically (that is, for a 2-level model with i nodes in the
first level and j child nodes per node in the first level so that the columns and rows of the precision matrix are
labelled B_11, ..., B_ij, B_1, ..., B_i, B) into a depth-first labelling (for example, for a 2-level model, B_11, B_12,
..., B_1J, B_1, B_21, ..., B_2J, B_2, ...)
}
\examples{
i <- 2
j <- 3
k <- 2
levels <- 3
depthfirst(i = i, j = j, k = k, levels = levels)
}
