% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillin3.R
\name{fillin3}
\alias{fillin3}
\title{centered_fillin3}
\usage{
fillin3(method, permute_method = "none", i, J, K, ...)
}
\arguments{
\item{method}{determines which linear algebra package to use}

\item{permute_method}{determines how the precision matrix is permuted before use with the chosen linear algebra package}

\item{i}{number of nodes at level 1}

\item{J}{vector specifying the number of children nodes in level 2 per node at level 1}

\item{K}{vector specifying the number of children nodes in level 3 per node at level 2}

\item{...}{for other params used in centered_precgen3}
}
\value{
the fill-in ratio with respect to the optimal fill-in, and the time taken for the cholesky decomposition
}
\description{
This function returns the ratio of the number of zeros obtained from the cholesky decomposition of the
specified precision matrix over the optimal cholesky fill-in for a general 3-level gaussian hierarchical model.
If the cholesky decomposition is optimal, the number returned is 1, and if it is sub-optimal the number returned
is greater than 1. The function also returns the time taken for the cholesky decomposition.
}
\details{
A choice of 2 packages is provided: spam and Matrix.
A choice of 3 permute methods is provided: no permutation, random permutation, and depth-first permutation.
}
\examples{
i <- 2
J <- c(1,2)
K <- c(1,2,3)
fillin3(method = 'matrix', permute_method = 'none', i = i, J=J, K=K)
}
